Improve textboxes (UI/UX)

+++ delete-list.php
@@ -18,8 +18,9 @@
         //SElect Database
         $db_select = mysqli_select_db($conn, DB_NAME) or die(mysqli_error());
         
+        $current_user_id = $_SESSION['user_id'];
         //Write the Query to DELETE List from DAtabase
-        $sql = "DELETE FROM tbl_lists WHERE list_id=$list_id";
+        $sql = "DELETE FROM tbl_lists WHERE list_id=$list_id AND user_id=$current_user_id";
         
         //Execute The Query
         $res = mysqli_query($conn, $sql);
@@ -27,6 +28,9 @@
         //Check whether the query executed successfully or not
         if($res==true)
         {
+            // Also delete associated tasks
+            mysqli_query($conn, "DELETE FROM tbl_tasks WHERE list_id=$list_id AND user_id=$current_user_id");
+            
             //Query Executed Successfully which means list is deleted successfully
             $_SESSION['delete'] = "List Deleted Successfully";
